Move test_read_csv_pragmatic input files
