+++ tables/test/input/test_read_csv_pragmatic/all.csv
