+++ tables/test/input/test_read_csv_pragmatic/auto_fixed.py
