+++ tables/test/input/test_read_csv_pragmatic/cols2.csv
