+++ tables/test/input/test_read_csv_pragmatic/ex0.csv
