+++ tables/test/input/test_read_csv_pragmatic/ex2.csv
