+++ tables/test/input/test_read_csv_pragmatic/ex3.csv
