+++ tables/test/input/test_read_csv_pragmatic/row1.csv
