+++ tables/test/input/test_read_csv_pragmatic/types2.csv
