+++ tables/test/input/test_read_csv_pragmatic/types3.csv
